OpenTransport optimization from Terje Norderhaug:
I have a potential improvement for the ACL-SOCKET-MCL.LISP module used in AEServe.

ACL-SOCKET-MCL scans all streams of the socket to determine which one has become established. However, the OpenTransport module of MCL establishes connections in the order their stream were created. It follows that only the earliest created stream has to be tested to determine whether a connection is established.