Fix error (?) in authorization pattern.
For some reason, the authorization pattern here specified `(:accept
"127.1")` even though a comment indicates that this is supposed to
allow from localhost.  Since localhost is 127.0.0.1, this pattern
didn't match, causing a large number of the tests to fail.

Since it's hard to see how this could *ever* have worked, I'm not at
all confident in this change.